+++ src/main/java/com/michal/kuc/liftctl/repositories/CarriageStatusRepository.java
@@ -4,6 +4,8 @@ import com.michal.kuc.liftctl.model.CarriageStatus;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface CarriageStatusRepository extends MongoRepository<CarriageStatus, Integer> {
+public interface CarriageStatusRepository extends MongoRepository<CarriageStatus, BigInteger> {
 }
